Added Javadocs for many classes and methods. Added a range JUnit test. Reduced PMD warnings.
