+++ core/src/main/java/gov/nist/secauto/metaschema/core/metapath/antlr/Metapath10ParserBase.java
@@ -31,10 +31,22 @@ import org.antlr.v4.runtime.TokenStream;
 
 public abstract class Metapath10ParserBase
     extends Parser {
+  /**
+   * Construct a new parser base.
+   *
+   * @param input
+   *          the input token stream
+   */
   protected Metapath10ParserBase(TokenStream input) {
     super(input);
   }
 
+  /**
+   * Check if functional call name does not include a keyword.
+   *
+   * @return {@code true} if the function call name is free of keywords, or
+   *         {@code false} otherwise
+   */
   protected boolean isFuncCall() {
     return !(getInputStream().LA(1) == Metapath10.KW_ARRAY
         || getInputStream().LA(1) == Metapath10.KW_ATTRIBUTE
